+++ GFramework.Cqrs.Tests/Cqrs/GeneratedRegistryNotification.cs
@@ -1,0 +1,8 @@
+using GFramework.Cqrs.Abstractions.Cqrs;
+
+namespace GFramework.Cqrs.Tests.Cqrs;
+
+/// <summary>
+///     用于验证生成注册器路径的通知消息。
+/// </summary>
+internal sealed record GeneratedRegistryNotification : INotification;
